store status
